+++ src/app/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import * as React from "react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import MemeList from "@/components/memes/MemeList";
