+++ src/app/topics/page.tsx
@@ -1,11 +1,9 @@
-"use client";
-
 import * as React from "react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { SortBy } from "@/services/meme/types";
 import { TopicSortBy } from "@/services/topic/types";
 import TopicList from "@/components/topic/TopicList";
-import { SearchInput } from "@/components/ui/search-input";
+// import { SearchInput } from "@/components/ui/search-input";
 
 const capitalizeFirstLetter = (str: string) =>
   str[0].toUpperCase() + str.slice(1);
