+++ src/app/u/[addr]/page.tsx
@@ -1,11 +1,13 @@
-"use client";
-
 import UserBaseInfo from "@/components/user/details/UserBaseInfo";
 import UserTabs from "@/components/user/details/UserTabs";
-import { useParams } from "next/navigation";
 
-export default function UserDetails() {
-  const params = useParams<{ addr: string }>();
+export default function UserDetails({
+  params,
+}: {
+  params: {
+    addr: string;
+  };
+}) {
   const addr = params.addr;
   return (
     <div className="flex flex-row justify-center items-start gap-6 max-sm:flex-col max-sm:gap-3">
