+++ src/services/user/types.ts
@@ -9,7 +9,7 @@ export type UserData = {
 
 export type UserLeaderboardData = {
   user: {
-    walletAddress: string;
+    walletAddress: Address;
     name?: string;
     avatar?: string;
   };
@@ -31,6 +31,9 @@ export type OwnedMemeData = {
   graduation?: MemeGraduation;
   referralReward?: {
     amount: number;
+    proof: any;
+    index: number;
+    root: any;
     isClaimed: boolean;
   };
 };
